+++ bos/core/b_device.c
@@ -326,7 +326,7 @@ int bDeviceCtl(uint32_t dev_no, uint8_t cmd, void *param)
 
 int bDeviceModifyHalIf(uint32_t dev_no, uint32_t offset, const uint8_t *pVal, uint8_t size)
 {
-    uint32_t halif_addr = 0;
+    uintptr_t halif_addr = 0;
     if (dev_no >= B_REG_DEV_NUMBER || pVal == NULL || size == 0)
     {
         return -1;
@@ -335,7 +335,7 @@ int bDeviceModifyHalIf(uint32_t dev_no, uint32_t offset, const uint8_t *pVal, ui
     {
         return -2;
     }
-    halif_addr = ((uint32_t)bDriverInterfaceTable[dev_no].hal_if) + offset;
+    halif_addr = ((uintptr_t)bDriverInterfaceTable[dev_no].hal_if) + offset;
 #if (defined(_HALIF_VARIABLE_ENABLE) && (_HALIF_VARIABLE_ENABLE == 1))
     memcpy((uint8_t *)halif_addr, pVal, size);
     return 0;
